# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MinRegistry management API",
    "description": "Management API of MinRegistry. All endpoints except this document require an admin session (GitHub OAuth); mutating requests must send `X-Requested-With: XMLHttpRequest`.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/repositories": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "Repositories, alphabetically, with a substring filter.",
        "operationId": "list_repositories",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Substring filter on the repository name.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (default 50, max 500).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryListResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/repositories/{id}": {
      "get": {
        "tags": [
          "repositories"
        ],
        "summary": "A repository with its tags and manifests (including referrers).",
        "operationId": "get_repository",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Repository id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "repositories"
        ],
        "summary": "Deletes a repository: tags, manifests, grants and uploads go; blobs are\nreclaimed by the next garbage collection.",
        "operationId": "delete_repository",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Repository id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/repositories/{id}/manifests/{digest}": {
      "delete": {
        "tags": [
          "repositories"
        ],
        "summary": "Deletes a manifest and every tag pointing at it.",
        "operationId": "delete_repository_manifest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Repository id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "digest",
            "in": "path",
            "description": "Manifest digest",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/repositories/{id}/permissions": {
      "get": {
        "tags": [
          "permissions"
        ],
        "summary": "Who may access a repository.",
        "operationId": "list_repository_permissions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Repository id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RepositoryPermissionSummary"
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/repositories/{id}/permissions/{principal_id}": {
      "put": {
        "tags": [
          "permissions"
        ],
        "summary": "Grants (or changes) a principal's level on a repository.",
        "operationId": "grant_permission",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Repository id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "principal_id",
            "in": "path",
            "description": "Principal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantPermissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryPermissionSummary"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "permissions"
        ],
        "summary": "Removes a principal's grant on a repository.",
        "operationId": "revoke_permission",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Repository id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "principal_id",
            "in": "path",
            "description": "Principal id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/repositories/{id}/tags/{tag}": {
      "delete": {
        "tags": [
          "repositories"
        ],
        "summary": "Removes a tag (the manifest stays).",
        "operationId": "delete_repository_tag",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Repository id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "path",
            "description": "Tag name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable code, e.g. `not_found`, `unauthenticated`, `csrf`."
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "GrantPermissionRequest": {
        "type": "object",
        "required": [
          "level"
        ],
        "properties": {
          "level": {
            "$ref": "#/components/schemas/PermissionLevel"
          }
        }
      },
      "ManifestSummary": {
        "type": "object",
        "required": [
          "digest",
          "media_type",
          "size",
          "platforms",
          "annotations",
          "tags",
          "created_at"
        ],
        "properties": {
          "digest": {
            "type": "string"
          },
          "media_type": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "artifact_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "subject_digest": {
            "type": [
              "string",
              "null"
            ],
            "description": "Set for referrers (signatures, SBOMs, …): the manifest they refer to."
          },
          "platforms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformSummary"
            }
          },
          "annotations": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "string"
          },
          "pushed_by": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PermissionLevel": {
        "type": "string",
        "enum": [
          "read",
          "write",
          "owner"
        ]
      },
      "PlatformSummary": {
        "type": "object",
        "required": [
          "os",
          "architecture"
        ],
        "properties": {
          "os": {
            "type": "string"
          },
          "architecture": {
            "type": "string"
          },
          "variant": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PrincipalKind": {
        "type": "string",
        "enum": [
          "github",
          "identity"
        ]
      },
...
      "PrincipalRefSummary": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/PrincipalKind"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "RepositoryDetail": {
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "tags",
          "manifests"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "created_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagSummary"
            }
          },
          "manifests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManifestSummary"
            }
          }
        }
      },
      "RepositoryListResponse": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RepositorySummary"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "RepositoryPermissionSummary": {
        "type": "object",
        "required": [
          "principal",
          "level",
          "granted_at"
        ],
        "properties": {
          "principal": {
            "$ref": "#/components/schemas/PrincipalRefSummary"
          },
          "level": {
            "$ref": "#/components/schemas/PermissionLevel"
          },
          "granted_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "granted_at": {
            "type": "string"
          }
        }
      },
...
      "RepositorySummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "tag_count",
          "manifest_count"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "created_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "tag_count": {
            "type": "integer",
            "format": "int64"
          },
          "manifest_count": {
            "type": "integer",
            "format": "int64"
          },
          "last_push_at": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TagSummary": {
        "type": "object",
        "required": [
          "name",
          "digest",
          "updated_at"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "digest": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "updated_by": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "minregistry_session"
      }
    }
  },
...
}